ajout d'une redirection automatique si les donnés client sont undefined

+++ Kanap/front/js/confirmation.js
@@ -1,7 +1,14 @@
-const id = new URL(window.location.href).searchParams.get("id");
+const id = new URL(window.location.href).searchParams.get('id');
 console.log(id);
 
 const orderId = document.getElementById('orderId');
-orderId.innerText = id;
+let p = document.querySelector('p')
 
-localStorage.clear();
+if (id ==='undefined') {
+        p.innerText = 'Erreur lors de la commande, veuillez remplir le formulaire entièrement!' 
+        setTimeout('window.location="cart.html"',6000)
+        
+}else{
+        orderId.innerText = id;
+        localStorage.clear();
+}
